requirements.txt
# requirements.txt
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.1
tenacity==8.2.3
streamlit==1.32.0
openai==1.12.0